pyproject.toml
[project]
name = "veda-devtools"
version = "0.1.0"
description = "A typed DSL (VedaLang) that compiles to VEDA tables for TIMES energy models"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "python-dotenv>=1.0",
    "tabulate>=0.9.0",
    "gitpython>=3.1.45",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "jinja2>=3.1.6",
    # xl2times dependencies
    "loguru>=0.7",
    "numpy>=1.26",
    "tqdm>=4.66",
    "more-itertools>=10.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
    "watchdog>=6.0.0",
    # LSP server dependencies
    "pygls>=1.3.0",
    "lsprotocol>=2024.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
vedalang = "vedalang.cli:main"
vedalang-dev = "tools.veda_dev.cli:main"

[tool.setuptools.packages.find]
include = ["tools*", "vedalang*", "xl2times*"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
# HTML templates contain long CSS/JS lines
"tools/veda_dev/sankey.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-s --tb=short"
